requirements.txt
appdirs==1.4.2
Babel==2.3.4
Django==1.10.5
django-countries==4.1
django-crispy-forms==1.6.1
django-datetime-widget==0.9.3
django-phonenumber-field==1.1.0
django-registration-redux==1.4
packaging==16.8
phonenumberslite==8.3.2
psycopg2==2.7
pyparsing==2.1.10
pytz==2016.10
six==1.10.0
